Corrigindo conflitos do [merge 773ed78](https://github.com/Is-It-Safe/mobile/pull/98/commits/773ed7872b8437c20572e0462353cfa0f75caf92)
