+++ lib/src/app/modules/location/presenter/pages/location_page.dart
@@ -38,210 +38,124 @@ class _LocationPageState extends ModularState<LocationPage, LocationBloC> {
   Widget build(BuildContext context) {
     final size = MediaQuery.of(context).size;
     return Scaffold(
-      floatingActionButton: FloatingActionButton(
-        onPressed: () async {
-          await Modular.to.pushNamed(
-            LocationModule.route + ReviewPage.route,
-            arguments: widget.location,
-          );
-        },
-        backgroundColor: SafeColors.buttonColors.secondary,
-        child: Icon(
-          Icons.add,
-          color: Colors.white,
-          size: size.width * .11,
+        floatingActionButton: FloatingActionButton(
+          onPressed: () async {
+            await Modular.to.pushNamed(
+              LocationModule.route + ReviewPage.route,
+              arguments: widget.location,
+            );
+          },
+          backgroundColor: SafeColors.buttonColors.secondary,
+          child: Icon(
+            Icons.add,
+            color: Colors.white,
+            size: size.width * .11,
+          ),
         ),
-      ),
-      appBar: SafeAppBar(
-        title: widget.location.name,
-        onTap: () {
-          Modular.to.pop();
-        },
-      ),
-      body: StreamBuilder<SafeEvent<LocationEntity>>(
-          stream: controller.locationByIDController.stream,
-          builder: (context, snapshot) {
-            if (snapshot.hasData) {
-              final loc = controller.locationById;
-              return SafeLayout(
-                snapshot: snapshot,
-                onCompleted: SingleChildScrollView(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Image.network(
-                        loc.imagePath ?? StringConstants.empty,
-                        width: size.width,
-                        fit: BoxFit.cover,
-                      ),
-                      SizedBox(height: size.height * .032),
-                      Container(
-                        margin: EdgeInsets.only(
-                          left: size.width * .08454,
-                          right: size.width * .273,
+        appBar: SafeAppBar(
+          title: widget.location.name,
+          onTap: () {
+            Modular.to.pop();
+          },
+        ),
+        body: StreamBuilder<SafeEvent<LocationEntity>>(
+            stream: controller.locationByIDController.stream,
+            builder: (context, snapshot) {
+              if (snapshot.hasData) {
+                final loc = controller.locationById;
+                return SafeLayout(
+                  snapshot: snapshot,
+                  onCompleted: SingleChildScrollView(
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Image.network(
+                          loc.imagePath ?? StringConstants.empty,
+                          width: size.width,
+                          fit: BoxFit.cover,
                         ),
-                        child: Column(
-                          children: [
-                            Text(
-                              loc.name,
-                              overflow: TextOverflow.clip,
-                              style: TextStyles.headline3().copyWith(
-                                fontSize: 22,
-                                fontWeight: FontWeight.w700,
+                        SizedBox(height: size.height * .032),
+                        Container(
+                          margin: EdgeInsets.only(
+                            left: size.width * .08454,
+                            right: size.width * .273,
+                          ),
+                          child: Column(
+                            children: [
+                              Text(
+                                loc.name,
+                                overflow: TextOverflow.clip,
+                                style: TextStyles.headline3().copyWith(
+                                  fontSize: 22,
+                                  fontWeight: FontWeight.w700,
+                                ),
                               ),
-                            ),
-                            const SizedBox(height: 4),
-                            Text(
-                              loc.address,
-                              overflow: TextOverflow.clip,
-                              style: TextStyles.caption().copyWith(
-                                fontWeight: FontWeight.w500,
+                              const SizedBox(height: 4),
+                              Text(
+                                loc.address,
+                                overflow: TextOverflow.clip,
+                                style: TextStyles.caption().copyWith(
+                                  fontWeight: FontWeight.w500,
+                                ),
                               ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      SizedBox(height: size.height * .047),
-                      SafeEmotionGrapic(
-                        grade: loc.averageGrade ?? 0,
-                        avaliationCount: loc.reviews!.length,
-                      ),
-                      SizedBox(height: size.height * .023),
-                      Container(
-                        margin: EdgeInsets.symmetric(
-                            horizontal: size.width * .08212),
-                        child: const SafeImpressionCard(
-                          type: SafeImpressionEnum.success,
-                        ),
-                      ),
-                      SizedBox(height: size.height * .035),
-                      Divider(
-                        color: SafeColors.generalColors.secondary,
-                        thickness: 1,
-                        indent: size.width * .0845,
-                        endIndent: size.width * .0845,
-                      ),
-                      SizedBox(height: size.height * .037),
-                      Container(
-                        margin: EdgeInsets.only(left: size.width * .084),
-                        child: Text(
-                          S.current.textWhatPeopleAreTalkingAboutThisPlace,
-                          style: TextStyles.subtitle2().copyWith(
-                            fontWeight: FontWeight.w700,
+                            ],
                           ),
                         ),
-                      ),
-                      SizedBox(height: size.height * .023),
-                      Container(
-                        margin: EdgeInsets.only(
-                          left: size.width * .07,
-                          right: size.width * .0845,
-                          bottom: size.height * .058,
+                        SizedBox(height: size.height * .047),
+                        SafeEmotionGrapic(
+                          grade: loc.averageGrade ?? 0,
+                          avaliationCount: loc.reviews!.length,
                         ),
-                        child: ListView.builder(
-                          shrinkWrap: true,
-                          physics: const NeverScrollableScrollPhysics(),
-                          itemCount: loc.reviews!.length,
-                          itemBuilder: (_, index) {
-                            return SafeReviewCard(
-                              review:
-                                  ReviewEntity.toEntity(loc.reviews![index]),
-                            );
-                          },
+                        SizedBox(height: size.height * .023),
+                        Container(
+                          margin: EdgeInsets.symmetric(
+                              horizontal: size.width * .08212),
+                          child: const SafeImpressionCard(
+                            type: SafeImpressionEnum.success,
+                          ),
                         ),
-                      ),
-                    ],
-                  ),
-                ),
-              );
-            }
-            return const SafeLoading();
-          }),
-          
-import 'package:is_it_safe_app/src/app/modules/location/presenter/bloc/location_bloc.dart';
-import 'package:is_it_safe_app/src/app/modules/location/presenter/widgets/location_mout_textfield.dart';
-import 'package:is_it_safe_app/src/app/modules/location/presenter/widgets/location_photo.dart';
-import 'package:is_it_safe_app/src/components/config/safe_event.dart';
-import 'package:is_it_safe_app/src/components/config/safe_layout.dart';
-import 'package:is_it_safe_app/src/components/widgets/safe_app_bar.dart';
-import 'package:is_it_safe_app/src/components/widgets/safe_button.dart';
-import 'package:is_it_safe_app/src/domain/entity/location_entity.dart';
-
-class SaveLocationPage extends StatefulWidget {
-  static const route = '/save_location/';
-
-  const SaveLocationPage({Key? key}) : super(key: key);
-
-  @override
-  State<SaveLocationPage> createState() => _SaveLocationPageState();
-}
-
-class _SaveLocationPageState
-    extends ModularState<SaveLocationPage, SaveLocationBloc> {
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: SafeAppBar(
-        title: S.current.textAddLocationSubTitle,
-      ),
-      body: StreamBuilder<SafeEvent<LocationEntity>>(
-        stream: controller.isSavingLocation.stream,
-        initialData: SafeEvent.initial(),
-        builder: (context, snapshot) {
-          return SafeLayout(
-            snapshot: snapshot,
-            context: context,
-            onCompleted: const SizedBox.shrink(),
-            onInitial: SingleChildScrollView(
-              child: Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 24.0,
-                  vertical: 8.0,
-                ),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    MountTextField(
-                      formKey: _formKey,
-                      controller: controller,
-                    ),
-                    ValueListenableBuilder<String?>(
-                      valueListenable: controller.imageNotifier,
-                      builder: (context, value, _) {
-                        return Padding(
-                          padding: const EdgeInsets.only(bottom: 48.0),
-                          child: LocalePhotoComponent(
-                            path: value,
-                            onTap: () async {
-                              controller.imageNotifier.value =
-                                  await controller.handleCameraTap();
+                        SizedBox(height: size.height * .035),
+                        Divider(
+                          color: SafeColors.generalColors.secondary,
+                          thickness: 1,
+                          indent: size.width * .0845,
+                          endIndent: size.width * .0845,
+                        ),
+                        SizedBox(height: size.height * .037),
+                        Container(
+                          margin: EdgeInsets.only(left: size.width * .084),
+                          child: Text(
+                            S.current.textWhatPeopleAreTalkingAboutThisPlace,
+                            style: TextStyles.subtitle2().copyWith(
+                              fontWeight: FontWeight.w700,
+                            ),
+                          ),
+                        ),
+                        SizedBox(height: size.height * .023),
+                        Container(
+                          margin: EdgeInsets.only(
+                            left: size.width * .07,
+                            right: size.width * .0845,
+                            bottom: size.height * .058,
+                          ),
+                          child: ListView.builder(
+                            shrinkWrap: true,
+                            physics: const NeverScrollableScrollPhysics(),
+                            itemCount: loc.reviews!.length,
+                            itemBuilder: (_, index) {
+                              return SafeReviewCard(
+                                review:
+                                    ReviewEntity.toEntity(loc.reviews![index]),
+                              );
                             },
                           ),
-                        );
-                      },
+                        ),
+                      ],
                     ),
-                    Center(
-                      child: SafeButton(
-                        title: S.current.textAddLocationConfirm,
-                        hasBackground: true,
-                        size: ButtonSize.large,
-                        onTap: () async {
-                          if (_formKey.currentState?.validate() ?? false) {
-                            await controller.sendNewLocation();
-                          }
-                        },
-                      ),
-                    ),
-                    const SizedBox(height: 26),
-                  ],
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
+                  ),
+                );
+              }
+              return const SafeLoading();
+            }));
   }
 }
