+++ lib/src/app/modules/location/review/presenter/pages/review_page.dart
@@ -61,7 +61,6 @@ class ReviewPageState extends ModularState<ReviewPage, ReviewBloc> {
             //     ).show(),
             //   );
             return SafeLayout(
-              context: context,
               snapshot: snapshot,
               onCompleted: Padding(
                 padding: const EdgeInsets.symmetric(
