+++ lib/src/components/config/safe_layout.dart
@@ -25,7 +25,6 @@ class SafeLayout extends StatelessWidget {
     this.onEmpty = const SizedBox.shrink(),
     this.onInitial = const SizedBox.shrink(),
     this.onError = const SizedBox.shrink(),
-    required BuildContext context,
   }) : super(key: key);
 
   @override
