+++ lib/src/domain/use_case/get_best_rated_locations_use_case.dart
@@ -13,8 +13,6 @@ class GetBestRatedLocationsUseCase extends SafeUseCase {
 
   Future<List<LocationEntity>> call({String? place}) async {
     final response = await _service.getBestRatedPlaces(place);
-    return response
-        .map((e) => LocationEntity.toEntityGetRatedPlaces(e))
-        .toList();
+    return response.map((e) => LocationEntity.toEntity(e)).toList();
   }
 }
